pyproject.toml
[tool.poetry]
name = "TRSS-Plugin"
version = "1.0.0"
description = "TRSS Yunzai Plugin"
authors = ["时雨🌌星空 <Time.Rain.Star.Sky@Gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.6.2"
torch = "^1.13.1"
scipy = "^1.9.3"
librosa = "^0.9.2"
unidecode = "^1.3.6"
phonemizer = "^3.2.1"
pypinyin = "^0.47.1"
pypinyin-dict = "^0.4.0"
jieba = "^0.42.1"
cython = "^0.29.32"
basicsr = "^1.4.2"
facexlib = "^0.2.5"
gfpgan = "^1.3.8"
numpy = "^1.23.5"
opencv-python = "^4.6.0.66"
pillow = "^9.3.0"
torchvision = "^0.14.1"
tqdm = "^4.64.1"
flask = "^2.2.2"
pycryptodome = "^3.16.0"
requests = "^2.28.1"

[[tool.poetry.source]]
name = "USTC"
url = "https://pypi.mirrors.ustc.edu.cn/simple"
default = true